- upped default max zoom to 20 (from 17)
- fixed/updated Google maps base url to support zooms 17+
- fixed occasional crash when inserting tile to pixmap cache (ImageManager)
- clearer separation of tile downloading and caching (ImageManager)
- additional tile servers available for OSM map adapter (OpenTopo, OpenCycle,Stamen)
- network request assumed time-outed are automatically retried (NetworkManager)
- tiles for disk-cache only are no longer needlessly decoded once downloaded
- failed downloads are now signalled (NetworkManager)
- Qt network replies are now deleted wjen no longer needed to keep the memory usage with a lot of downloads in check
  (easily took GB of mem with 10000s tiles downloaded)
- progress indicator widget can be hidden (QMapControl)
- zoom level changes are now signaled (QMapControl)

+++ QMapControl/src/QMapControl/MapAdapterWMS.cpp
@@ -141,8 +141,8 @@ namespace qmapcontrol
         QUrlQuery url_query(getBaseUrl());
 
         // Calculate the number of coordinates per tile.
-        const int coord_per_tile_x = 360.0 / projection::get().tilesX(controller_zoom);
-        const int coord_per_tile_y = 180.0 / projection::get().tilesY(controller_zoom);
+        const int coord_per_tile_x = int(360.0 / projection::get().tilesX(controller_zoom));
+        const int coord_per_tile_y = int(180.0 / projection::get().tilesY(controller_zoom));
 
         // Set BBOX (x1,y1,x2,y2).
         url_query.removeQueryItem("BBOX");
